LexHex S-Record: use the required data field size for some highlighting Equal to how it is already implemeted for the Intel HEX lexer.
Specific record types do not use a data field, the checksum field
directly follows the address field then.
In these cases now the byte count field is also marked as wrong if its
content is not what is to be expected for this record type. Also the
checksum field is now fixed after the address field instead of allowing
data field in between if the byte count is too big.

Also simplify this code path for IHex to be equal to the Srec code.