Support for static compilations on Windows
When you compile the code on Win32, the code needs to be exported/imported accross DLL boundaries. However - this is not meeded when you build the code as a static library. 

On such cases - just defining that macro to be empty is enough (like on Unix). Seems to work for MinGW on Qt6. I assume this will work for MSVC Qt55/Qt6.

Nothing changed on Linux/OSX (this is guarded by a `#ifdef WIN32`.